+++ src/main/java/controllers/prospects/DeleteProspectsController.java
@@ -23,7 +23,7 @@ public final class DeleteProspectsController implements ICommand {
         String urlSuite = Security.estConnecte(request, jsp);
 
         if (jsp.equals(urlSuite)) {
-            Long id = Long.parseLong(request.getParameter("id"));
+            Integer id = Integer.parseInt(request.getParameter("id"));
             Prospect prospect = prospectService.findById(id);
             
             if (prospect != null) {
